# docs/swagger.yml
swagger: "2.0"
info:
  description: "Swagger file for mess-app API specification"
  version: "1.0.0"
  title: "Mess Split Bill"
  contact:
    email: "tshihad9@gmail.com"
  license:
    name: "Apache 2.0"
    url: "http://www.apache.org/licenses/LICENSE-2.0.html"
    
tags:
- name: "user"
  description: "User operations"
schemes:
- "https"
- "http"
paths:
  /{version}/user/{user_name}:
    get:
      tags:
      - "user"
      summary: "Get user"
      description: "user_name provides in url"
      produces:
      - "application/json"
      parameters:
      - $ref: '#/parameters/Version'
      - name: user_name
        in: "path"
        description: "user name of the user"
        required: true
        type: "string"
      responses:
        200:
          description: "ok"
          schema:
            $ref: "#/responses/UserResponse"
        404:
          description: "Invalid user name"
    delete:
      tags:
      - user
      summary: "Delete user" 
      description: "Delete user by user_name"
      produces: 
      - "application/json"
      parameters: 
      - $ref: '#/parameters/Version'
      - name: user_name
        in: path
        description: "user name of the user"
        required: true
        type: "string"
      responses:
        202:
          description: OK
        404:
          description: resource not found

  /{version}/user:
    post:
      tags:
      - "user"
      summary: "Add new user"
      description: "3 Type of user(1. Admin, 2. Collegues, 3.Cook)"
      operationId: "PostUser"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - $ref: '#/parameters/Version'
      - in: "body"
        name: "body"
        description: "user object needed to added in users table"
        required: true
        schema:
          $ref: "#/definitions/UserPayload"
      responses:
        201:
          description: "ok" 
          schema:
            $ref: "#/responses/UserResponse"
        405:
          description: "Invalid input"

definitions:
  UserPayload:
    type: "object"
    required:
    - "user_name"
    - "email"
    - "user_type"
    - "password"
    properties:
      user_name:
        type: "string"
        example: shihad
      email:
        type: "string"
        example: tshihad9@gmail.com
      user_type:
        type: "string"
        description: "Only 3 values accepted (admin,collegues,cook)"
        example: "admin"
        pattern: admin|cook|collegues
      password:
        type: string
        example: password123
        description: "Password for user"
  resp:
    type: "object"
    properties:
      status:
        type: string
        pattern: OK|ERROR
        example: ERROR

  UserResponsePayload:
    type: "object"
    required:
    - "user_name"
    - "email"
    - "user_type"
    properties:
      user_name:
        type: "string"
        example: shihad
      email:
        type: "string"
        example: tshihad9@gmail.com
      user_type:
        type: "string"
        description: "Only 3 values accepted (admin,collegues,cook)"
        example: "admin"
        pattern: admin|cook|collegues
responses:
  UserResponse:
    description: Client endpoint response
    schema:
      allOf:
        - $ref: "#/definitions/resp"
        - $ref: "#/definitions/UserResponsePayload"

parameters:
  Version:
    name: version
    description: default v1
    in: path
    required: true
    type: string
    pattern: 'v[0-9]'